refactor settings-tab.tst.ts into smaller files
